Cold attack damages the enemy when hit
Added a few lines into spear.cs.
Declared variable damage.
Added OnCollision2D() to detect whether the spear is hit to any object, and damages if it is not the player.

Created a new script named "temp_Enemy_Health", which is temporarily added into "EnemyDraft" object. It prints its current health every 0.5 seconds on console.

Set spear prefab's original position to (0, 0) which was previously around (-1.4, 0). I believe it should be originated to the origin.